Move redis exporter to a package
