+++ exporter/redis_test.go
@@ -1,4 +1,4 @@
-package main
+package exporter
 
 /*
   to run the tests with redis running on anything but localhost:6379 use
